backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1
slowapi==0.1.9
minio>=7.2.0
python-dotenv==1.0.1
boto3>=1.34.0